Drops DataTable, which truncates precision at hundredths.
