+++ src/nv2adbg/_program_inputs_viewer.py
@@ -4,14 +4,15 @@ from typing import Optional
 
 from textual.app import ComposeResult
 from textual.containers import Horizontal
+from textual.containers import VerticalScroll
 from textual.css import query
 from textual.reactive import reactive
 from textual.widgets import ContentSwitcher
-from textual.widgets import DataTable
 from textual.widgets import Static
 
-from nv2adbg import simulator
 from nv2adbg._error_message import _CenteredErrorMessage
+from nv2adbg._register_view import _RegisterSetPanel
+from nv2adbg.simulator import Context
 
 
 class _ProgramInputsViewer(Static):
@@ -26,57 +27,63 @@ class _ProgramInputsViewer(Static):
     _ProgramInputsViewer DataTable {
         height: 1fr;
     }
+
+    #inputs-table .registersetpanel--register {
+        color: $success;
+        text-style: bold;
+    }
+
+    #constants-table .registersetpanel--border {
+        color: $warning;
+    }
+
+    #constants-table .registersetpanel--register {
+        color: $success;
+        text-style: bold;
+    }
     """
 
     _active_content = reactive("no-content", init=False)
 
     def __init__(self):
         super().__init__()
-        self._context: Optional[simulator.Context] = None
+        self._context: Optional[Context] = None
 
-        self._inputs_table = DataTable(id="inputs-table")
-        self._inputs_table.add_columns("Input", "x", "y", "z", "w")
-        self._inputs_table.cursor_type = "row"
-        self._inputs_table.zebra_stripes = True
+        self._inputs_table = _RegisterSetPanel("Inputs", id="inputs-table")
+        self._constants_table = _RegisterSetPanel("Constants", id="constants-table")
 
-        self._constants_table = DataTable(id="constants-table")
-        self._constants_table.add_columns("Constant", "x", "y", "z", "w")
-        self._constants_table.cursor_type = "row"
-        self._constants_table.zebra_stripes = True
-
-    def set_context(self, context: simulator.Context):
+    def set_context(self, context: Context):
         self._context = context
-        self._populate_table()
+        self._populate()
         self.update()
 
     def on_mount(self) -> None:
-        self._populate_table()
+        self._populate()
 
     def compose(self) -> ComposeResult:
         with ContentSwitcher(initial="no-content"):
             with Horizontal(id="content"):
-                yield self._inputs_table
-                yield self._constants_table
+                with VerticalScroll():
+                    yield self._inputs_table
+                with VerticalScroll():
+                    yield self._constants_table
 
             yield _CenteredErrorMessage(
                 "No input context available, load data via the File menu.",
                 id="no-content",
             )
 
-    def _populate_table(self):
+    def _populate(self):
         try:
-            self._inputs_table.clear()
-            self._constants_table.clear()
-
             if self._context:
-                for register in self._context.inputs:
-                    self._inputs_table.add_row(
-                        register.name, register.x, register.y, register.z, register.w
-                    )
-                for register in self._context.constants:
-                    self._constants_table.add_row(
-                        register.name, register.x, register.y, register.z, register.w
-                    )
+                self._inputs_table.set_registers(self._context.inputs)
+
+                def rename_constant_register(name: str) -> str:
+                    return f"c[{name[1:]}]"
+
+                self._constants_table.set_registers(
+                    self._context.constants, rename_constant_register
+                )
                 self._active_content = "content"
             else:
                 self._active_content = "no-content"
